hw1/requirements.txt
torch
torchvision
matplotlib
scikit-learn
pandas
pillow
